requirements.txt
langchain
langchain-core
langgraph
langchain-community
langchain-google-genai
google-generativeai
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
arxiv
requests
python-dotenv
tavily-python